feat: 增强文件处理和本地化支持
添加了文件下载和反序列化功能，改进了配置管理。
- 在 `FileHelper.cs` 中，新增 `DownloadFileAsync` 方法文档注释，并确保 `ExtractLanguagePackage` 方法的流参数不为空。
- 在 `JsonHelper.cs` 中，新增 `Deserialize<T>` 方法以支持 JSON 反序列化。
- 删除项目文件 `LLC_MOD_Toolbox.csproj` 中的 `Assets\I18n\` 文件夹项。
- 在 `MainWindow.xaml.cs` 中，添加对 `SevenZip` 的引用，注释掉与进度计时器相关的代码，并新增 `StartGreytestButtonClick` 方法以处理灰度测试逻辑。
- 在 `Config.cs` 中，将 `GamePath` 属性修改为可空类型，并添加静态方法 `Create` 从 URL 创建配置。
- 在 `AutoInstallerViewModel.cs` 中，新增 `testToken` 属性以存储配置中的 Token。
- 在 `SettingsViewModel.cs` 中，将 `LimbusCompanyPath` 属性修改为可空类型，并在删除 BepInEx 时添加空值检查。
- 在 `zh-CN.xaml` 中，添加资源字典以支持自动安装的本地化字符串。
- 新增 `ApiLoadingTextService` 和 `FileLoadingTextService` 服务类，分别用于通过 API 和随机选择加载文本。
- 定义 `ILoadingTextService` 接口，包含异步获取加载文本的方法。
