+++ Assets/I18n/zh-CN.xaml
@@ -1,0 +1,6 @@
+<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:sys="clr-namespace:System;assembly=mscorlib">
+    <sys:String x:Key="自动安装">自动安装</sys:String>
+</ResourceDictionary>
