+++ LLC_MOD_Toolbox.csproj
@@ -73,8 +73,4 @@
     </Compile>
   </ItemGroup>
 
-  <ItemGroup>
-    <Folder Include="Assets\I18n\" />
-  </ItemGroup>
-
 </Project>
